Revert "refactor: move backend files into a module so we can use it in notebooks too"
This reverts commit 55a49a86a10e602370cddf997532a2245cb037dc.
